+++ election-server/src/util.rs
@@ -1,4 +1,4 @@
-use anyhow::{ensure, Context};
+use anyhow::Context;
 use concordium_governance_committee_election::ElectionConfig;
 use concordium_rust_sdk::{
     contract_client::{ContractClient, ViewError},
@@ -8,7 +8,6 @@ use concordium_rust_sdk::{
 };
 use eg::ballot::BallotEncrypted;
 use serde::Serialize;
-use std::{fs::File, path::Path};
 use tonic::transport::ClientTlsConfig;
 
 pub const REGISTER_VOTES_RECEIVE: &str = "election.registerVotes";
@@ -40,24 +39,6 @@ pub struct VotingWeightDelegation {
 
 pub enum ElectionContractMarker {}
 pub type ElectionContract = ContractClient<ElectionContractMarker>;
-
-/// Verify the digest of `file` matches the expected `checksum`.
-pub fn verify_checksum(file: &Path, expected_checksum: [u8; 32]) -> anyhow::Result<()> {
-    use sha2::Digest;
-
-    let mut hasher = sha2::Sha256::new();
-    let mut source =
-        File::open(file).with_context(|| format!("Unable to open file at path {file:?}"))?;
-
-    std::io::copy(&mut source, &mut hasher)
-        .with_context(|| format!("Could not digest file at location {file:?}"))?;
-    let computed_hash: [u8; 32] = hasher.finalize().into();
-    ensure!(
-        computed_hash == expected_checksum,
-        "Hash of file did not match checksum"
-    );
-    Ok(())
-}
 
 /// Creates a [`v2::Client`] from the [`v2::Endpoint`], enabling TLS and setting
 /// connection and request timeouts