+++ apps/guardian/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+- Improve error message when requests to get remote resources fail
+
 ## 0.1.2
 
 - Updates concordium rust dependencies